Configure the Router to send the user to the correct page from the URL
The router is now configured so it sends users to the blog posts if they go directly to those and also to the projects if they enter the correct url.

Added the .DS_Store to the gitignore to prevent OSX directory files making it into the repository.

Created the basic configuration file for the Project where I will be declaring consts

[+] Configuartion File
[+] Routing 
[+] Prevented OSX files making it into the repository